03_hello_bot/sync_hello_ag/pyproject.toml
[project]
name = "sync-hello-ag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.7",
    "python-dotenv>=1.1.0",
]
